tweaks
- updated random number generator (speed!)
- wrapped `sinf()` / `cosf()` into `sincosf()`
- 4kB less stack in `make_analysis_window()`
- tweaks in the synthesizer

+++ inc/codec2_internal.h
@@ -127,7 +127,7 @@ typedef struct nlp_t
 
 typedef struct codec2_t
 {
-    unsigned long next_rn;
+    uint32_t next_rn;
 
     float w[M_PITCH];
     float W[FFT_ENC];
